+++ src/test/java/esa/sen2vm/DirectLocGridTest.java
@@ -90,7 +90,7 @@ public class DirectLocGridTest
     @Test
     public void createGrid()
     {
-        float step = 4.5f;
+        double step = 4.5f;
         int startLine = 0;
         int startPixel = 0;
         int sizeLine = 2304;
@@ -103,18 +103,19 @@ public class DirectLocGridTest
         assertEquals(dirGrid.getLineOffsetGranule(200), -3.75);
         assertEquals(dirGrid.getLineOffsetGranule(210), -0.25);
 
-        ArrayList<Float> pixels = dirGrid.getGridPixels();
+        ArrayList<Double> pixels = dirGrid.getGridPixels();
         assertEquals(pixels.size(), 95.0);
         assertEquals(pixels.get(0), -1.75f);
         assertEquals(pixels.get(1), pixels.get(0) + step);
         assertEquals(pixels.get(dirGrid.getGridPixels().size() - 1), 421.25f);
 
-        ArrayList<Float> lines = dirGrid.getGridLines();
+        ArrayList<Double> lines = dirGrid.getGridLines();
         assertEquals(lines.size(), 513);
         assertEquals(lines.get(0), -1.75f);
         assertEquals(lines.get(1), lines.get(0) + step);
         assertEquals(lines.get(dirGrid.getGridLines().size() - 1), 2302.25f);
     }
+
 
 
 }
